fix: move util to message/message.go to avoid import cycle
